Allow changing status of closed patches in last CF
Without this, once a patch was closed (rejected/committed etc),
it was no longer possible to undo that and reopen it.

We now allow this to be done in the very last commitfest a patch
is on. We cannot change it on a previous commitfest, as that could
end up with the patch being open in two CFs at the same time, but
there is no reason it shouldn't be doable in the last one.