Rename "returned with feedback"->"move to next cf", add new "returned with feedback"
Returned with Feedback will now actually close the patch and *not* move it to
the next CF. Since it used to do that, all existing Returned with Feedback patches
are changed to Moved to next CF which does the same thing that Returned with
Feedback used to do. And of course, Moved to next CF is a new option available for
each patch from now on.